# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "Codebase API Example documentation",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/goals/v1": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List goals with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Goals"
                ],
                "summary": "List goals",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc|desc)",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wrapper.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/schema.ResponseGoalItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wrapper.JSONResult"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new goal (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Goals"
                ],
                "summary": "Create a new goal",
                "parameters": [
                    {
                        "description": "Create goal",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.RequestGoalCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wrapper.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.ResponseGoalCreate"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wrapper.JSONResult"
                        }
                    }
                }
            }
        },
        "/goals/v1/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a single goal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Goals"
                ],
                "summary": "Get goal by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Goal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wrapper.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.ResponseGoalGet"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wrapper.JSONResult"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a goal (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Goals"
                ],
                "summary": "Delete a goal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Goal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wrapper.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.ResponseGoalDelete"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wrapper.JSONResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "contract.StatusCode": {
            "type": "string",
            "enum": [
                "000000",
                "000001",
                "000002",
                "000011",
                "000012",
                "000013",
                "000014"
            ],
            "x-enum-varnames": [
                "StatusCodeSuccess",
                "StatusCodeBindingFailed",
                "StatusCodeValidationFailed",
                "StatusCodeUnauthorized",
                "StatusCodeUserOrPasswordInvalid",
                "StatusCodeInternalServerError",
                "StatusCodeSequenceError"
            ]
        },
        "middleware.AuthUserData": {
            "type": "object",
            "properties": {
                "role": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "schema.RequestGoalCreate": {
            "type": "object",
            "required": [
                "match_id",
                "minute_scored",
                "player_id"
            ],
            "properties": {
                "authUserData": {
                    "$ref": "#/definitions/middleware.AuthUserData"
                },
                "match_id": {
                    "type": "string"
                },
                "minute_scored": {
                    "type": "integer",
                    "minimum": 0
                },
                "player_id": {
                    "type": "string"
                }
            }
        },
...
        "schema.ResponseGoalCreate": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "schema.ResponseGoalDelete": {
            "type": "object"
        },
        "schema.ResponseGoalGet": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "match_id": {
                    "type": "string"
                },
                "minute_scored": {
                    "type": "integer"
                },
                "player_id": {
                    "type": "string"
                }
            }
        },
        "schema.ResponseGoalItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "match_id": {
                    "type": "string"
                },
                "minute_scored": {
                    "type": "integer"
                },
                "player_id": {
                    "type": "string"
                }
            }
        },
...
        "wrapper.JSONResult": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/wrapper.PaginationMeta"
                },
                "statusCode": {
                    "$ref": "#/definitions/contract.StatusCode"
                }
            }
        },
        "wrapper.PaginationMeta": {
            "type": "object",
            "properties": {
                "metadata": {},
                "page": {
                    "type": "integer"
                },
                "totalData": {
                    "type": "integer"
                },
                "totalDataOnPage": {
                    "type": "integer"
                },
                "totalPage": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}